+++ src/util/Quaternion.java
@@ -3,11 +3,11 @@ package util;
 public class Quaternion {
 	// COOLbeans... this is a spherical linear interpolation between two
 	// Quaternions
-	public static Quaternion slerp(Quaternion a, Quaternion b, float t) {
-		final float epsilon = 0.00001f;
-		float flip = 1;
+	public static Quaternion slerp(Quaternion a, Quaternion b, double t) {
+		final double epsilon = 0.00001f;
+		double flip = 1;
 
-		float cosine = a.w * b.w + a.x * b.x + a.y * b.y + a.z * b.z;
+		double cosine = a.w * b.w + a.x * b.x + a.y * b.y + a.z * b.z;
 
 		if (cosine < 0) {
 			cosine = -cosine;
@@ -17,21 +17,21 @@ public class Quaternion {
 		if ((1 - cosine) < epsilon)
 			return a.scale(1 - t).add(b.scale(t * flip));
 		else {
-			float theta = (float) Math.acos(cosine);
-			float sine = (float) Math.sin(theta);
-			float beta = (float) Math.sin((1 - t) * theta) / sine;
-			float alpha = (float) Math.sin(t * theta) / sine * flip;
+			double theta = (double) Math.acos(cosine);
+			double sine = (double) Math.sin(theta);
+			double beta = (double) Math.sin((1 - t) * theta) / sine;
+			double alpha = (double) Math.sin(t * theta) / sine * flip;
 
 			return a.scale(beta).add(b.scale(alpha));
 		}
 	}
 
-	public float w, x, y, z;
+	public double w, x, y, z;
 
 	public Quaternion() {
 	}
 
-	public Quaternion(float w, float x, float y, float z) {
+	public Quaternion(double w, double x, double y, double z) {
 		this.w = w;
 		this.x = x;
 		this.y = y;
@@ -42,8 +42,8 @@ public class Quaternion {
 		return new Quaternion(w + o.w, x + o.x, y + o.y, z + o.z);
 	}
 
-	public float length() {
-		return (float) Math.sqrt(w * w + x * x + y * y + z * z);
+	public double length() {
+		return (double) Math.sqrt(w * w + x * x + y * y + z * z);
 	}
 
 	public Quaternion multiply(Quaternion o) {
@@ -55,7 +55,7 @@ public class Quaternion {
 	}
 
 	public Quaternion normalize() {
-		float length = this.length();
+		double length = this.length();
 		Quaternion quaternion = new Quaternion();
 		quaternion.w = w / length;
 		quaternion.x = x / length;
@@ -64,23 +64,23 @@ public class Quaternion {
 		return quaternion;
 	}
 
-	public Quaternion scale(float o) {
+	public Quaternion scale(double o) {
 		return new Quaternion(o * w, o * x, o * y, o * z);
 	}
 
 	public Matrix toMatrix() {
-		float fTx = 2.0f * x;
-		float fTy = 2.0f * y;
-		float fTz = 2.0f * z;
-		float fTwx = fTx * w;
-		float fTwy = fTy * w;
-		float fTwz = fTz * w;
-		float fTxx = fTx * x;
-		float fTxy = fTy * x;
-		float fTxz = fTz * x;
-		float fTyy = fTy * y;
-		float fTyz = fTz * y;
-		float fTzz = fTz * z;
+		double fTx = 2.0f * x;
+		double fTy = 2.0f * y;
+		double fTz = 2.0f * z;
+		double fTwx = fTx * w;
+		double fTwy = fTy * w;
+		double fTwz = fTz * w;
+		double fTxx = fTx * x;
+		double fTxy = fTy * x;
+		double fTxz = fTz * x;
+		double fTyy = fTy * y;
+		double fTyz = fTz * y;
+		double fTzz = fTz * z;
 
 		return new Matrix(
 				1.0f - (fTyy + fTzz),
